requirements.txt
fastapi
uvicorn
orjson
sqlalchemy
python-jose[cryptography]
requests